Added RuneManager and refactored code to improve OO implementation
Now the RuneManager will take care of the runes balances of the rune
collectors. The rune collectors is a list of Actors who can keep runes.

Also refactored some areas of the code which were using downcasting.
This was fixed by changing the parameter type of the Collectable and
Consumable interface to Actor instead of Player.

There is still some refactoring to be done in PurchaseAction and
SellAction, there is downcasting from Item to WeaponItem. The first
thing that came to mind to fix this was to create a new similar class
called SellWeaponAction and PurchaseWeaponAction, and use those
accordingly based on the type of trade being made. This is because
MerchantKale only sells WeaponItems, there might be a future Trader who
sells Items. But the issue with this is that if MerchantKale was to sell
items we would have to implement some extra functionality to do use the
PurchaseAction and SellAction instead of the PurchaseWeaponAction and
SellWeaponAction.