Implemented FingerReaderEnia with TradeAction. Also some downcasting fixes.
Implemented FingerReaderEnia by using the same functionality as it's parent class
Trader. But I added trading for FingerReaderEnia since they must be able
to trade the RemembranceOfTheGrafted for either AxeOfGodrick or GraftedDragon.

I also did some refactoring for some downcasting issues in PurchaseAction,
SellAction and TradeAction, an issue related to WeaponItem and Item. This
was fixed by adding a seperate HashMap for weapons in the Trader abstract
class.