Finishing touches. Fixed Rune dropping.
Fixed issue where the rune was not being dropped because, Rune's were
not portable, so I set Rune's to portable and overrid the getPickUpAction
method from the superclass to prevent the Player from being able to pickup
Runes and keep them in the inventory. I want it so that Runes are only to
be consumed from the ground when the Player is standing on top of it, I
don't want the Player to be able to pickup the Rune.